Refactor: python, cts: Move RemoteBasic into the _cts module.
No code changes.
